AB#335 Move code to the original chart
