+++ client/src/pages/organizations/Laydown.tsx
@@ -1,7 +1,7 @@
 import AppContext from "components/AppContext"
 import EditAdministratingPositionsModal from "components/EditAdministratingPositionsModal"
 import Fieldset from "components/Fieldset"
-import OrganizationalChart from "components/graphs/OrganizationalChart2"
+import OrganizationalChart from "components/graphs/OrganizationalChart"
 import LinkTo from "components/LinkTo"
 import Model from "components/Model"
 import PositionTable from "components/PositionTable"
